Add keep option to replace original files with compressed files
When this option is disabled, this is a backwards incompatible change, but
for apps supporting internet explorer (most new apps and certainly most apps
wanting to use brotli) its the best default.
Also, this is the same default used by ember-cli-deploy-gzip, which it kind
of brings this addon closer to being a "better replacement" for it.